move InvalidSection component to its own file
