+++ src/components/Section/Package/InvalidForm.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { MemoryRouter } from 'react-router'
 import { mount } from 'enzyme'
-import InvalidForm, { InvalidSection } from './InvalidForm'
+import InvalidForm from './InvalidForm'
 
 describe('The Invalid form submission component', () => {
   it('renders errors', () => {
@@ -27,22 +27,5 @@ describe('The Invalid form submission component', () => {
     }
     const component = mount(<MemoryRouter><InvalidForm {...expected} /></MemoryRouter>)
     expect(component.find('.invalid-form').length).toBe(1)
-  })
-
-  it('renders invalid section', () => {
-    const expected = {
-      onClick: () => {},
-      mark: {
-        section: {
-          title: 'Cool'
-        },
-        subsections: [{
-          name: 'Full name',
-          url: 'name'
-        }]
-      }
-    }
-    const component = mount(<MemoryRouter><InvalidSection {...expected} /></MemoryRouter>)
-    expect(component.find('.error-messages').length).toBe(1)
   })
 })
